YAML: promote BlockScalarUtils helpers onto Yaml.Scalar
Move getBody() and withBody(String) / withBody(String, int) from the
internal BlockScalarUtils onto Yaml.Scalar itself as instance methods, per
review discussion. The internal utility class and its dedicated test are
deleted; the four CRLF-boundary unit tests move to ScalarTest under a
@Nested BlockScalarBody block.

Consumers updated to call scalar.getBody() / scalar.withBody(...) instead
of the static helpers: ChangePropertyValue, ChangeValue, YamlValue,
YamlReference. No behavior change in any consumer.

The Javadoc pointer on Yaml.Scalar.value now references withBody(String)
directly instead of the internal utility.
